# swagger.yaml
...
openapi: "3.0.3"
info:
  version: development
  title: Ticketing Application
  description: "API documentation for **Ticketing Application**"
servers:
  - url:  http://localhost:8080
    description: "Development"
...
paths:
...
  /ticket/update:
    post:
      tags:
        - ticket
      summary: update ticket
      operationId: updateTicket
      requestBody:
        description: Request body to update ticket
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                      id: 55
                      event: "konser musik"
                      schedule: "2022-05-01 19:30 +07"
                      quota: 34
                      price: 250000
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTicketRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      meta:
                        message: Update ticket data
                        code: 200
                        status: success
                      data:
                        id: 55
                        event: "konser musik"
                        schedule: "2022-05-01 19:30 +07"
                        quota: 34
                        price: 250000
  /ticket/delete:
    post:
      tags:
        - ticket
      summary: delete ticket
      operationId: deleteTicket
      requestBody:
        description: Request body to delete ticket
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTicketRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  1
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTicketRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      {
                          "meta": {
                              "message": "Delete ticket data",
                              "code": 200,
                              "status": "success"
                          },
                          "data": null
                      }                                      
...
  /ticket:
    post:
      tags:
        - ticket
      summary: get ticket by id
      operationId: getTicketById
      requestBody:
        description: Request body to get ticket by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTicketRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  55
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTicketRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                    meta:
                      message: Get ticket data
                      code: 200
                      status: success
                    data:
                      id: 54
                      event: "kkonser musik"
                      schedule: "2022-06-01 19:30 +07"
                      quota: 28
                      price: 250000
...
  /checkout/delete:
    post:
      tags:
        - checkout
      summary: delete checkout
      operationId: deleteCheckout
      requestBody:
        description: Request body to delete checkout
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCheckoutRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  54
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCheckoutRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      {
                          "meta": {
                              "message": "Delete checkout data",
                              "code": 200,
                              "status": "success"
                          },
                          "data": null
                      }                                      
...
  /checkout:
    post:
      tags:
        - checkout
      summary: get checkout by checkout id
      operationId: getCheckoutById
      requestBody:
        description: Request body to get checkout by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetCheckoutRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  54
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCheckoutRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      meta:
                        message: Get checkout data
                        code: 200
                        status: success
                      data:
                        id: 54
                        user_id: 66
                        ticket_id: 51
                        name: "anton"
                        event: "kkonser"
                        quantity: 2
                        timeout: "2022-04-13 03:38 +07"
                        price: 250000
                        total_amount: 500000
...
  /transaction/delete:
    post:
      tags:
        - transaction
      summary: delete transaction
      operationId: deleteTransaction
      requestBody:
        description: Request body to delete transaction
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTransactionRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  65
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteTransactionRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      {
                          "meta": {
                              "message": "Delete transaction data",
                              "code": 200,
                              "status": "success"
                          },
                          "data": null
                      }                                      
...
  /transaction:
    post:
      tags:
        - transaction
      summary: get transaction by transaction id
      operationId: getTransactionById
      requestBody:
        description: Request body to get transaction by id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTransactionRq'
            examples:
              ScenarioA:
                summary: "Scenario A"
                value:
                  id:  54
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTransactionRs'
              examples:
                ScenarioA:
                  summary: "Scenario OK"
                  value:
                      meta:
                        message: Get transaction data
                        code: 200
                        status: success
                      data:
                        - id: 65
                          user_id: 66
                          ticket_id: 54
                          name: anton
                          event: kkonser musik
                          quantity: 2
                          status: PAID
                          price: 250000
                          total_amount: 500000
components:
  schemas:
...
    UpdateTicketRq:
      type: object
      required:
        - id
        - event
        - schedule
        - price
        - quota
      properties:
        id:
          type: integer
        event:
          type: string
        schedule:
          type: string
          pattern: "2022-02-01 19:30 +07"
        price:
          type: integer
        quota:
          type: integer
    UpdateTicketRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  event:
                    type: string
                  schedule:
                    type: string
                    pattern: "2022-02-01 19:30 +07"
                  price:
                    type: integer
                  quota:
                    type: integer
    DeleteTicketRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    DeleteTicketRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
...
    GetTicketRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    GetTicketRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  event:
                    type: string
                  schedule:
                    type: string
                    pattern: "2022-02-01 19:30 +07"
                  price:
                    type: integer
                  quota:
                    type: integer                        
...
    DeleteCheckoutRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    DeleteCheckoutRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
...
    GetCheckoutRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    GetCheckoutRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  ticket_id:
                    type: integer                    
                  name:
                    type: string                    
                  event:
                    type: string
                  quantity:
                    type: integer  
                  timeout:
                    type: string
                    pattern: "2022-02-01 19:30 +07"
                  price:
                    type: integer
                  total_amount:
                    type: integer                                   
...
    DeleteTransactionRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    DeleteTransactionRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
...
    GetTransactionRq:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    GetTransactionRs:
        allOf: 
          - $ref: "#/components/schemas/ApiResponse"
          - type: "object"
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  ticket_id:
                    type: integer                    
                  name:
                    type: string                    
                  event:
                    type: string
                  quantity:
                    type: integer  
                  status:
                    type: string
                  price:
                    type: integer
                  total_amount:
                    type: integer                  
    ApiResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            message:
              type: string
            code:
              type: integer
            status:
              type: string